refactor: Move methods out of main to utils
Rename `_format_content` to better (and non private) name `format_default_values`
