+++ completor/main.py
@@ -18,7 +18,14 @@ from completor.get_version import get_version
 from completor.launch_args_parser import get_parser
 from completor.logger import handle_error_messages, logger
 from completor.read_casefile import ReadCasefile
-from completor.utils import abort, clean_file_line, clean_file_lines
+from completor.utils import (
+    abort,
+    clean_file_lines,
+    clean_raw_data,
+    find_keyword_data,
+    find_well_keyword_data,
+    format_default_values,
+)
 from completor.wells import Well
 
 
@@ -143,7 +150,7 @@ def process_content(line_number: int, clean_lines: dict[int, str]) -> tuple[list
         line_number += 1
         if line_number in clean_lines:
             content += clean_lines[line_number]
-    content = _format_content(content)
+    content = format_default_values(content)
     return content, line_number
 
 
@@ -182,44 +189,35 @@ def create(
     meaningful_data: ScheduleData = {}
 
     try:
-        keyword = Keywords.WELL_SPECIFICATION
-        chunks = find_keyword_data(keyword, schedule)
-        for chunk in chunks:
-            clean_data = clean_raw_data(chunk, keyword)
+        # Find the old data for each of the four main keywords.
+        for chunk in find_keyword_data(Keywords.WELL_SPECIFICATION, schedule):
+            clean_data = clean_raw_data(chunk, Keywords.WELL_SPECIFICATION)
             meaningful_data = read_schedule.set_welspecs(meaningful_data, clean_data)
 
-        keyword = Keywords.COMPLETION_DATA
-        chunks = find_keyword_data(keyword, schedule)
-        for chunk in chunks:
-            clean_data = clean_raw_data(chunk, keyword)
+        for chunk in find_keyword_data(Keywords.COMPLETION_DATA, schedule):
+            clean_data = clean_raw_data(chunk, Keywords.COMPLETION_DATA)
             meaningful_data = read_schedule.set_compdat(meaningful_data, clean_data)
 
-        keyword = Keywords.WELL_SEGMENTS
-        chunks = find_keyword_data(keyword, schedule)
-        for chunk in chunks:
-            clean_data = clean_raw_data(chunk, keyword)
+        for chunk in find_keyword_data(Keywords.WELL_SEGMENTS, schedule):
+            clean_data = clean_raw_data(chunk, Keywords.WELL_SEGMENTS)
             meaningful_data = read_schedule.set_welsegs(meaningful_data, clean_data)
 
-        keyword = Keywords.COMPLETION_SEGMENTS
-        chunks = find_keyword_data(keyword, schedule)
-        for chunk in chunks:
-            clean_data = clean_raw_data(chunk, keyword)
+        for chunk in find_keyword_data(Keywords.COMPLETION_SEGMENTS, schedule):
+            clean_data = clean_raw_data(chunk, Keywords.COMPLETION_SEGMENTS)
             meaningful_data = read_schedule.set_compsegs(meaningful_data, clean_data)
 
         for i, well_name in tqdm(enumerate(active_wells.tolist()), total=len(active_wells)):
-            # Note: Important to run this check before creating wells as it fixes potential problems with cases.
-            case.check_input(well_name, meaningful_data)
             well = Well(well_name, i, case, meaningful_data[well_name])
             compdat, welsegs, compsegs, bonus = create_output.format_output(well, case, figure_name)
 
             for keyword in [Keywords.COMPLETION_SEGMENTS, Keywords.WELL_SEGMENTS, Keywords.COMPLETION_DATA]:
-                tmp_data = find_well_keyword_data(well_name, keyword, schedule)
-                old_data = str("\n".join(tmp_data))
+                # TODO: Move to utils? parsing?
+                old_data = find_well_keyword_data(well_name, keyword, schedule)
+                if not old_data:
+                    raise CompletorError(
+                        "Could not find the unmodified data in original schedule file. Please contact the team!"
+                    )
                 try:
-                    if not old_data:
-                        raise CompletorError(
-                            "Could not find the unmodified data in original schedule file. Please contact the team!"
-                        )
                     # Check that nothing is lost.
                     schedule.index(old_data)
                 except ValueError:
@@ -317,134 +315,6 @@ def _get_well_name(schedule_lines: dict[int, str], i: int) -> str:
     j = np.where(keys == i)[0][0]
     next_line = schedule_lines[int(keys[j + 1])]
     return next_line.split()[0]
-
-
-def _format_content(text: str) -> list[list[str]]:
-    """Format the data-records and resolve the repeat-mechanism.
-
-    E.g. 3* == 1* 1* 1*, 3*250 == 250 250 250.
-
-    Args:
-        text: A chunk data-record.
-
-    Returns:
-        Expanded values.
-    """
-    chunk = re.split(r"\s+/", text)[:-1]
-    expanded_data = []
-    for line in chunk:
-        new_record = ""
-        for record in line.split():
-            if not record[0].isdigit():
-                new_record += record + " "
-                continue
-            if "*" not in record:
-                new_record += record + " "
-                continue
-
-            # Handle repeats like 3* or 3*250.
-            multiplier, number = record.split("*")
-            new_record += f"{number if number else '1*'} " * int(multiplier)
-        if new_record:
-            expanded_data.append(new_record.split())
-    return expanded_data
-
-
-def find_keyword_data(keyword: str, text: str) -> list[str]:
-    """Finds the common pattern for the four keywords thats needed.
-
-    Args:
-        keyword: Current keyword.
-        text: The whole text to find matches in.
-
-    Returns:
-        The matches if any.
-
-    """
-    # Finds keyword followed by two slashes.
-    # Matches any characters followed by a newline, non-greedily, to allow for comments within the data.
-    # Matches new line followed by a single (can have leading whitespace) slash.
-    pattern = rf"^{keyword}(?:.*\n)*?\s*\/"
-    return re.findall(pattern, text, re.MULTILINE)
-
-
-def clean_raw_data(raw_record: str, keyword: str) -> list[list[str]]:
-    """Parse the record and clean its content.
-
-    Args:
-        raw_record: Raw data taken straight from schedule.
-        keyword: The current keyword.
-
-    Returns:
-        The contents of the keyword, cleaned.
-    """
-    record = re.split(rf"{keyword}\n", raw_record)
-    if len(record) != 2:
-        raise CompletorError(f"Something went wrong when reading keyword '{keyword}' from schedule:\n{raw_record}")
-    # Strip keyword and last line.
-    raw_content = record[1].splitlines()
-    if raw_content[-1].strip().startswith("/"):
-        raw_content = raw_content[:-1]
-
-    clean_content = []
-    for line in raw_content:
-        clean_line = clean_file_line(line, remove_quotation_marks=True)
-        if clean_line:
-            clean_content.append(_format_content(clean_line)[0])
-    return clean_content
-
-
-def find_well_keyword_data(well: str, keyword: str, text: str) -> list[str]:
-    """Find the data associated with keyword and well name, include leading comments.
-
-    Args:
-        well: Well name.
-        keyword: Keyword to search for.
-        text: Raw text to look for matches in.
-
-    Returns:
-        The correct match given keyword and well name.
-    """
-    matches = find_keyword_data(keyword, text)
-
-    lines = []
-    for match in matches:
-        if re.search(well, match) is None:
-            continue
-
-        matchlines = match.splitlines()
-        once = False
-        for i, line in enumerate(matchlines):
-            if not line:
-                # Allow empty lines in the middle of a record.
-                if once:
-                    lines.append(line)
-                continue
-            if well in line.split()[0]:
-                if keyword in [Keywords.WELL_SEGMENTS, Keywords.COMPLETION_SEGMENTS]:
-                    # These keywords should just be the entire match as they never contain more than one well.
-                    return matchlines
-                if not once:
-                    once = True
-                    # Remove contiguous comments above the first line by looking backwards,
-                    # adding it to the replaceable text match.
-                    comments = []
-                    for prev_line in matchlines[i - 1 :: -1]:
-                        if not prev_line.strip().startswith("--") or not prev_line:
-                            break
-                        comments.append(prev_line)
-                    lines += sorted(comments, reverse=True)
-                # else:
-                lines.append(line)
-            elif not once:
-                continue
-            # All following comments inside data.
-            elif line.strip().startswith("--"):
-                lines.append(line)
-            else:
-                break
-
-    return lines
 
 
 if __name__ == "__main__":
