+++ completor/utils.py
@@ -11,6 +11,7 @@ import numpy.typing as npt
 import pandas as pd
 
 from completor.constants import Content, Headers, Keywords
+from completor.exceptions import CompletorError
 from completor.logger import logger
 
 
@@ -260,3 +261,130 @@ def check_width_lines(result: str, limit: int) -> list[tuple[int, str]]:
         if len(cleaned_line) > limit:
             too_long_lines.append((line_index, lines[line_index]))
     return too_long_lines
+
+
+def format_default_values(text: str) -> list[list[str]]:
+    """Format the data-records and resolve the repeat-mechanism.
+
+    E.g. 3* == 1* 1* 1*, 3*250 == 250 250 250.
+
+    Args:
+        text: A chunk data-record.
+
+    Returns:
+        Expanded values.
+    """
+    chunk = re.split(r"\s+/", text)[:-1]
+    expanded_data = []
+    for line in chunk:
+        new_record = ""
+        for record in line.split():
+            if not record[0].isdigit():
+                new_record += record + " "
+                continue
+            if "*" not in record:
+                new_record += record + " "
+                continue
+
+            # Handle repeats like 3* or 3*250.
+            multiplier, number = record.split("*")
+            new_record += f"{number if number else '1*'} " * int(multiplier)
+        if new_record:
+            expanded_data.append(new_record.split())
+    return expanded_data
+
+
+def find_keyword_data(keyword: str, text: str) -> list[str]:
+    """Finds the common pattern for the four keywords thats needed.
+
+    Args:
+        keyword: Current keyword.
+        text: The whole text to find matches in.
+
+    Returns:
+        The matches if any.
+
+    """
+    # Finds keyword followed by two slashes.
+    # Matches any characters followed by a newline, non-greedily, to allow for comments within the data.
+    # Matches new line followed by a single (can have leading whitespace) slash.
+    pattern = rf"^{keyword}(?:.*\n)*?\s*\/"
+    return re.findall(pattern, text, re.MULTILINE)
+
+
+def clean_raw_data(raw_record: str, keyword: str) -> list[list[str]]:
+    """Parse the record and clean its content.
+
+    Args:
+        raw_record: Raw data taken straight from schedule.
+        keyword: The current keyword.
+
+    Returns:
+        The contents of the keyword, cleaned.
+    """
+    record = re.split(rf"{keyword}\n", raw_record)
+    if len(record) != 2:
+        raise CompletorError(f"Something went wrong when reading keyword '{keyword}' from schedule:\n{raw_record}")
+    # Strip keyword and last line.
+    raw_content = record[1].splitlines()
+    if raw_content[-1].strip().startswith("/"):
+        raw_content = raw_content[:-1]
+
+    clean_content = []
+    for line in raw_content:
+        clean_line = clean_file_line(line, remove_quotation_marks=True)
+        if clean_line:
+            clean_content.append(format_default_values(clean_line)[0])
+    return clean_content
+
+
+def find_well_keyword_data(well: str, keyword: str, text: str) -> str:
+    """Find the data associated with keyword and well name, include leading comments.
+
+    Args:
+        well: Well name.
+        keyword: Keyword to search for.
+        text: Raw text to look for matches in.
+
+    Returns:
+        The correct match given keyword and well name.
+    """
+    matches = find_keyword_data(keyword, text)
+
+    lines = []
+    for match in matches:
+        if re.search(well, match) is None:
+            continue
+
+        matchlines = match.splitlines()
+        once = False
+        for i, line in enumerate(matchlines):
+            if not line:
+                # Allow empty lines in the middle of a record.
+                if once:
+                    lines.append(line)
+                continue
+            if well in line.split()[0]:
+                if keyword in [Keywords.WELL_SEGMENTS, Keywords.COMPLETION_SEGMENTS]:
+                    # These keywords should just be the entire match as they never contain more than one well.
+                    return match
+                if not once:
+                    once = True
+                    # Remove contiguous comments above the first line by looking backwards,
+                    # adding it to the replaceable text match.
+                    comments = []
+                    for prev_line in matchlines[i - 1 :: -1]:
+                        if not prev_line.strip().startswith("--") or not prev_line:
+                            break
+                        comments.append(prev_line)
+                    lines += sorted(comments, reverse=True)
+                lines.append(line)
+            elif not once:
+                continue
+            # All following comments inside data.
+            elif line.strip().startswith("--"):
+                lines.append(line)
+            else:
+                break
+
+    return str("\n".join(lines))
